CONTA/requirements.txt
numpy>=1.16.3
cython
imageio>=2.5.0
scikit-image>=0.15.0
chainercv>=0.12.0
git+https://github.com/lucasb-eyer/pydensecrf.git
matplotlib
pyyaml
scipy
tqdm
addict
black
joblib
omegaconf
torchnet